+++ atoml/parser.py
@@ -371,7 +371,8 @@ class Parser:
                     raise self.parse_error(UnexpectedCharError, "=")
                 else:
                     found_equals = True
-            pass
+        if not found_equals:
+            raise self.parse_error(UnexpectedCharError, self._current)
 
         if not key.sep:
             key.sep = self.extract()